refactor: extract special-attributes/condition form section
