+++ src/kleinanzeigen_bot/resources/translations.de.yaml
@@ -156,26 +156,6 @@ kleinanzeigen_bot/__init__.py:
     "DONE: updated %s": "FERTIG: %s aktualisiert"
     "ad": "Anzeige"
 
-  _set_condition:
-    "Condition value [%s] is deprecated; update your config to [%s].": "Zustandswert [%s] ist veraltet; aktualisieren Sie die Konfiguration auf [%s]."
-    "Unable to close condition dialog!": "Kann den Dialog für Artikelzustand nicht schließen!"
-    "Unable to select condition [%s]": "Zustand [%s] konnte nicht ausgewählt werden"
-    "Failed to set attribute '%s'": "Fehler beim Setzen des Attributs '%s'"
-    "No condition radio matched values %(values)s": "Keine Zustand-Auswahl passt zu den Werten %(values)s"
-
-  _set_special_attributes:
-    "Found %i special attributes": "%i spezielle Attribute gefunden"
-    "Setting special attribute [%s] to [%s]...": "Setze spezielles Attribut [%s] auf [%s]..."
-    "Successfully set attribute field [%s] to [%s]...": "Attributfeld [%s] erfolgreich auf [%s] gesetzt..."
-    "Failed to set attribute '%s'": "Fehler beim Setzen des Attributs '%s'"
-    "Special attribute '%s' is not available for the selected category. Skipping.": "Spezielles Attribut '%s' ist für die ausgewählte Kategorie nicht verfügbar. Überspringe."
-    "Attribute field '%s' seems to be a select...": "Attributfeld '%s' scheint ein Auswahlfeld zu sein..."
-    "Failed to set attribute field '%s' via known input types.": "Fehler beim Setzen des Attributfelds '%s' über bekannte Eingabetypen."
-    "Attribute field '%s' seems to be a checkbox...": "Attributfeld '%s' scheint eine Checkbox zu sein..."
-    "Attribute field '%s' seems to be a text input...": "Attributfeld '%s' scheint ein Texteingabefeld zu sein..."
-    "Attribute field '%s' seems to be a Combobox (i.e. text input with filtering dropdown)...": "Attributfeld '%s' scheint eine Combobox zu sein (d.h. Texteingabefeld mit Dropdown-Filter)..."
-    "Condition dialog not available, falling back to generic attribute handler for [%s]...": "Zustandsdialog nicht verfügbar, falle auf generischen Attribut-Handler für [%s] zurück..."
-
   run:
     "DONE: No configuration errors found.": "FERTIG: Keine Konfigurationsfehler gefunden."
     "DONE: No active ads found.": "FERTIG: Keine aktiven Anzeigen gefunden."
@@ -283,6 +263,26 @@ kleinanzeigen_bot/publishing_form.py:
     "shipping_options must be provided": "shipping_options müssen angegeben werden!"
     "Unknown shipping option(s), please refer to the documentation/README: %s": "Unbekannte Versandoption(en), bitte lies die Dokumentation/das README: %s"
     "You can only specify shipping options for one package size!": "Du kannst nur Versandoptionen für eine Paketgröße angeben!"
+
+  _set_condition:
+    "Condition value [%s] is deprecated; update your config to [%s].": "Zustandswert [%s] ist veraltet; aktualisieren Sie die Konfiguration auf [%s]."
+    "Unable to close condition dialog!": "Kann den Dialog für Artikelzustand nicht schließen!"
+    "Unable to select condition [%s]": "Zustand [%s] konnte nicht ausgewählt werden"
+    "Failed to set attribute '%s'": "Fehler beim Setzen des Attributs '%s'"
+    "No condition radio matched values %(values)s": "Keine Zustand-Auswahl passt zu den Werten %(values)s"
+
+  set_special_attributes:
+    "Found %i special attributes": "%i spezielle Attribute gefunden"
+    "Setting special attribute [%s] to [%s]...": "Setze spezielles Attribut [%s] auf [%s]..."
+    "Successfully set attribute field [%s] to [%s]...": "Attributfeld [%s] erfolgreich auf [%s] gesetzt..."
+    "Failed to set attribute '%s'": "Fehler beim Setzen des Attributs '%s'"
+    "Special attribute '%s' is not available for the selected category. Skipping.": "Spezielles Attribut '%s' ist für die ausgewählte Kategorie nicht verfügbar. Überspringe."
+    "Attribute field '%s' seems to be a select...": "Attributfeld '%s' scheint ein Auswahlfeld zu sein..."
+    "Failed to set attribute field '%s' via known input types.": "Fehler beim Setzen des Attributfelds '%s' über bekannte Eingabetypen."
+    "Attribute field '%s' seems to be a checkbox...": "Attributfeld '%s' scheint eine Checkbox zu sein..."
+    "Attribute field '%s' seems to be a text input...": "Attributfeld '%s' scheint ein Texteingabefeld zu sein..."
+    "Attribute field '%s' seems to be a Combobox (i.e. text input with filtering dropdown)...": "Attributfeld '%s' scheint eine Combobox zu sein (d.h. Texteingabefeld mit Dropdown-Filter)..."
+    "Condition dialog not available, falling back to generic attribute handler for [%s]...": "Zustandsdialog nicht verfügbar, falle auf generischen Attribut-Handler für [%s] zurück..."
 
 #################################################
 kleinanzeigen_bot/publishing_persistence.py:
